requirement.txt
ansicon==1.89.0
ashlar==1.14.1
asteval==0.9.31
blessed==1.20.0
Bottleneck==1.3.5
certifi==2022.12.7
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==2.0.0
colorama==0.4.6
cycler==0.11.0
Cython==3.0.6
cytoolz==0.12.0
dask==2021.10.0
docopt==0.6.2
einops==0.6.1
exceptiongroup==1.2.0
filelock==3.12.2
flit_core==3.6.0
fonttools==4.38.0
fsspec==2022.11.0
future==0.18.3
g2o-python==0.0.12
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
imglyb==2.1.0
importlib-metadata==6.7.0
iniconfig==2.0.0
jgo==1.0.5
jinxed==1.2.1
joblib==1.3.2
JPype1==1.4.1
kiwisolver==1.4.5
labeling==0.1.12
lmfit==1.2.2
locket==1.0.0
m2stitch==0.2.2
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
networkx==2.6.3
numexpr==2.8.4
numpy==1.21.6
opencv-python==4.8.1.78
packaging==22.0
pandas==1.3.5
pandas-stubs==1.2.0.62
partd==1.2.0
Pillow==9.5.0
pip==22.3.1
pipreqs==0.4.13
pluggy==1.2.0
ply==3.11
psutil==5.9.0
pyimagej==1.2.1
pyjnius==1.6.1
pyparsing==3.1.1
pytest==7.4.3
pytest-datadir==1.4.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.31.0
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.7.3
scyjava==1.9.1
setuptools==65.6.3
SimpleITK==2.0.0
sip==6.6.2
six==1.16.0
threadpoolctl==3.1.0
tifffile==2020.10.1
timm==0.6.11
toml==0.10.2
toolz==0.12.0
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tornado==6.2
tqdm==4.66.1
typing_extensions==4.4.0
uncertainties==3.1.7
urllib3==2.0.7
wcwidth==0.2.12
wheel==0.37.1
wincertstore==0.2
xarray==0.20.1
yarg==0.1.10
zipp==3.15.0
